+++ heylogs-ext-forgejo/src/main/java/nbbrd/heylogs/ext/forgejo/Forgejo.java
@@ -1,31 +1,27 @@
 package nbbrd.heylogs.ext.forgejo;
 
-import lombok.NonNull;
 import nbbrd.design.DirectImpl;
 import nbbrd.heylogs.spi.Forge;
 import nbbrd.heylogs.spi.ForgeSupport;
 import nbbrd.service.ServiceProvider;
 
-import java.net.URL;
-import java.util.Arrays;
+import static nbbrd.heylogs.spi.ForgeRefType.*;
 
 @DirectImpl
 @ServiceProvider
 public final class Forgejo implements Forge {
 
-    static final String ID = "forgejo";
-
     @lombok.experimental.Delegate
     private final Forge delegate = ForgeSupport
             .builder()
-            .id(ID)
+            .id("forgejo")
             .moduleId("forgejo")
             .name("Forgejo")
             .compareLinkFactory(ForgejoCompareLink::parse)
-            .linkPredicate(Forgejo::isForgejoHost)
+            .knownHostPredicate(ForgeSupport.onHostContaining("codeberg"))
+            .parser(COMMIT, ForgejoCommitLink::parse, ForgejoCommitRef::parse)
+            .parser(ISSUE, ForgejoIssueLink::parse, ForgejoIssueRef::parse)
+            .parser(REQUEST, ForgejoRequestLink::parse, ForgejoRequestRef::parse)
+            .parser(MENTION, ForgejoMentionLink::parse, ForgejoMentionRef::parse)
             .build();
-
-    static boolean isForgejoHost(@NonNull URL expected) {
-        return Arrays.asList(expected.getHost().split("\\.", -1)).contains("codeberg");
-    }
 }