+++ heylogs-ext-gitlab/src/main/java/nbbrd/heylogs/ext/gitlab/GitLab.java
@@ -1,31 +1,27 @@
 package nbbrd.heylogs.ext.gitlab;
 
-import lombok.NonNull;
 import nbbrd.design.DirectImpl;
 import nbbrd.heylogs.spi.Forge;
 import nbbrd.heylogs.spi.ForgeSupport;
 import nbbrd.service.ServiceProvider;
 
-import java.net.URL;
-import java.util.Arrays;
+import static nbbrd.heylogs.spi.ForgeRefType.*;
 
 @DirectImpl
 @ServiceProvider
 public final class GitLab implements Forge {
 
-    public static final String ID = "gitlab";
-
     @lombok.experimental.Delegate
     private final Forge delegate = ForgeSupport
             .builder()
-            .id(ID)
+            .id("gitlab")
             .moduleId("gitlab")
             .name("GitLab")
             .compareLinkFactory(GitLabCompareLink::parse)
-            .linkPredicate(GitLab::isGitLabHost)
+            .knownHostPredicate(ForgeSupport.onHostContaining("gitlab"))
+            .parser(COMMIT, GitLabCommitLink::parse, GitLabCommitRef::parse)
+            .parser(ISSUE, GitLabIssueLink::parse, GitLabIssueRef::parse)
+            .parser(REQUEST, GitLabRequestLink::parse, GitLabRequestRef::parse)
+            .parser(MENTION, GitLabMentionLink::parse, GitLabMentionRef::parse)
             .build();
-
-    static boolean isGitLabHost(@NonNull URL expected) {
-        return Arrays.asList(expected.getHost().split("\\.", -1)).contains("gitlab");
-    }
 }